move simpleop to the mixin
